Entropycraft V0.1 (Full Version)
This is the full content of version 0.1 of entropycraft.

(Sorry for not having multiple detailed commits, this was a personal
experiment originally that I was asked to make a full release of. Future
versions will be composed of multiple commits as progress is made, or in
oter words I do know how to use git and will be using it properly in
the future here. If you were hoping for information on why something was
done a specific way... sorry if there is not a comment in the soure...)